zvariant

Module as_value

Utilities to (de)serialize a value as a zvariant::Value.

This is mainly useful for constructing a strongly-typed API for dealing with dictionaries containing string keys and variant values (a{sv} in D-Bus language) See the relevant FAQ entry in our book for more details and examples.

Modules

optional

Utilities to (de)serialize an optional value as a zvariant::Value.

Structs

Deserialize

A wrapper to deserialize a value to T: Type + serde::Deserialize.

Serialize

A wrapper to serialize T: Type + serde::Serialize as a value.

Functions

deserialize

Deserialize a value as a zvariant::Value.

serialize

Serialize a value as a zvariant::Value.