Enum zvariant::Value

pub enum Value<'a> {

    U8(u8),
    Bool(bool),
    I16(i16),
    U16(u16),
    I32(i32),
    U32(u32),
    I64(i64),
    U64(u64),
    F64(f64),
    Str(Str<'a>),
    Signature(Signature<'a>),
    ObjectPath(ObjectPath<'a>),
    Value(Box<Value<'a>>),
    Array(Array<'a>),
    Dict(Dict<'a, 'a>),
    Structure(Structure<'a>),
    Fd(Fd),

}

A generic container, in the form of an enum that holds exactly one value of any of the other types.

Note that this type corresponds to the VARIANT data type defined by the D-Bus specification and as such, its encoding is not the same as that of the enclosed value.

Examples

use std::convert::TryFrom;
use zvariant::{from_slice, to_bytes, EncodingContext, Value};

// Create a Value from an i16
let v = Value::new(i16::max_value());

// Encode it
let ctxt = EncodingContext::<byteorder::LE>::new_dbus(0);
let encoding = to_bytes(ctxt, &v).unwrap();

// Decode it back
let v: Value = from_slice(&encoding, ctxt).unwrap();

// Check everything is as expected
assert_eq!(i16::try_from(&v).unwrap(), i16::max_value());

Now let’s try a more complicated example:

use std::convert::TryFrom;
use zvariant::{from_slice, to_bytes, EncodingContext};
use zvariant::{Structure, Value, Str};

// Create a Value from a tuple this time
let v = Value::new((i16::max_value(), "hello", true));

// Same drill as previous example
let ctxt = EncodingContext::<byteorder::LE>::new_dbus(0);
let encoding = to_bytes(ctxt, &v).unwrap();
let v: Value = from_slice(&encoding, ctxt).unwrap();

// Check everything is as expected
let s = Structure::try_from(v).unwrap();
assert_eq!(
    <(i16, Str, bool)>::try_from(s).unwrap(),
    (i16::max_value(), Str::from("hello"), true),
);

Variants

U8(u8)

Bool(bool)

I16(i16)

U16(u16)

I32(i32)

U32(u32)

I64(i64)

U64(u64)

F64(f64)

Str(Str<'a>)

Signature(Signature<'a>)

ObjectPath(ObjectPath<'a>)

Value(Box<Value<'a>>)

Array(Array<'a>)

Dict(Dict<'a, 'a>)

Structure(Structure<'a>)

Fd(Fd)

Implementations

impl<'a> Value<'a>

pub fn new<T>(value: T) -> Self
where T: Into<Self> + DynamicType,

Make a Value for a given value.

In general, you can use Into trait on basic types, except when you explicitly need to wrap Value itself, in which case this constructor comes handy.

Examples

use zvariant::Value;

let s = Value::new("hello");
let u: Value = 51.into();
assert_ne!(s, u);

pub fn to_owned(&self) -> OwnedValue

Create an owned version of self.

Ideally, we should implement std::borrow::ToOwned trait for Value, but that’s implemented generically for us through impl<T: Clone> ToOwned for T and it’s not what we need/want.

pub fn value_signature(&self) -> Signature<'_>

Get the signature of the enclosed value.

pub fn downcast<T>(self) -> Option<T>
where T: TryFrom<Value<'a>> + ?Sized,

Try to get the underlying type T.

Note that TryFrom<Value> is implemented for various types, and it’s usually best to use that instead. However, in generic code where you also want to unwrap Value::Value, you should use this function (because TryFrom<Value> can not be implemented for Value itself as From<Value> is implicitly implemented for Value).

Examples

use std::convert::TryFrom;
use zvariant::{Result, Value};

fn value_vec_to_type_vec<'a, T>(values: Vec<Value<'a>>) -> Result<Vec<T>>
where
    T: TryFrom<Value<'a>>,
{
    let mut res = vec![];
    for value in values.into_iter() {
        res.push(value.downcast().unwrap());
    }

    Ok(res)
}

// Let's try u32 values first
let v = vec![Value::U32(42), Value::U32(43)];
let v = value_vec_to_type_vec::<u32>(v).unwrap();
assert_eq!(v[0], 42);
assert_eq!(v[1], 43);

// Now try Value values
let v = vec![Value::new(Value::U32(42)), Value::new(Value::U32(43))];
let v = value_vec_to_type_vec::<Value>(v).unwrap();
assert_eq!(v[0], Value::U32(42));
assert_eq!(v[1], Value::U32(43));

pub fn downcast_ref<T>(&'a self) -> Option<&'a T>
where T: ?Sized, &'a T: TryFrom<&'a Value<'a>>,

Try to get a reference to the underlying type T.

Same as downcast except it doesn’t consume self and get a reference to the underlying value.

Examples

use std::convert::TryFrom;
use zvariant::{Result, Value};

fn value_vec_to_type_vec<'a, T>(values: &'a Vec<Value<'a>>) -> Result<Vec<&'a T>>
where
    &'a T: TryFrom<&'a Value<'a>>,
{
    let mut res = vec![];
    for value in values.into_iter() {
        res.push(value.downcast_ref().unwrap());
    }

    Ok(res)
}

// Let's try u32 values first
let v = vec![Value::U32(42), Value::U32(43)];
let v = value_vec_to_type_vec::<u32>(&v).unwrap();
assert_eq!(*v[0], 42);
assert_eq!(*v[1], 43);

// Now try Value values
let v = vec![Value::new(Value::U32(42)), Value::new(Value::U32(43))];
let v = value_vec_to_type_vec::<Value>(&v).unwrap();
assert_eq!(*v[0], Value::U32(42));
assert_eq!(*v[1], Value::U32(43));