Function zvariant::from_slice_with_seed

pub fn from_slice_with_seed<'d, B, S>(
    bytes: &'d [u8],
    ctxt: EncodingContext<B>,
    seed: S
) -> Result<S::Value>where
    B: ByteOrder,
    S: DeserializeSeed<'d> + DynamicType,

Deserialize T from a given slice of bytes containing file descriptor indices, using the given seed.

Please note that actual file descriptors are not part of the encoding and need to be transferred via an out-of-band platform specific mechanism. The encoding only contain the indices of the file descriptors and hence the reason, caller must pass a slice of file descriptors.