zvariant

Struct DynamicTuple

pub struct DynamicTuple<T>(pub T);

A helper type to serialize or deserialize a tuple whose elements implement DynamicType but not Type.

This is required because tuples already have an implementation of DynamicType via the blanket implementation of DynamicType where T: Type, but that results in a bound of Type on each element, which is stronger than needed for serializing.

Tuple Fields
0: T