Struct zvariant::Fd

pub struct Fd(/* private fields */);

A RawFd wrapper.

See also OwnedFd if you need a wrapper that takes ownership of the file.

We wrap the RawFd type so that we can implement Serialize and Deserialize for it. File descriptors are serialized in a special way and you need to use specific serializer and deserializer API when file descriptors are or could be involved.