zvariant

Struct Optional

pub struct Optional<T>(/* private fields */);

An optional value.

Since D-Bus doesn’t have the concept of nullability, it uses a special value (typically the default value) as the null value. For example this signal uses empty strings for null values. Serde has built-in support for Option but unfortunately that doesn’t work for us. Hence the need for this type.

The serialization and deserialization of Optional relies on NoneValue implementation of the underlying type.

Examples

use zvariant::{serialized::Context, Optional, to_bytes, LE};

// `Null` case.
let ctxt = Context::new_dbus(LE, 0);
let s = Optional::<&str>::default();
let encoded = to_bytes(ctxt, &s).unwrap();
assert_eq!(encoded.bytes(), &[0, 0, 0, 0, 0]);
let s: Optional<&str> = encoded.deserialize().unwrap().0;
assert_eq!(*s, None);

// `Some` case.
let s = Optional::from(Some("hello"));
let encoded = to_bytes(ctxt, &s).unwrap();
assert_eq!(encoded.len(), 10);
// The first byte is the length of the string in Little-Endian format.
assert_eq!(encoded[0], 5);
let s: Optional<&str> = encoded.deserialize().unwrap().0;
assert_eq!(*s, Some("hello"));