zvariant

Trait DynamicType

pub trait DynamicType {
    // Required method
    fn dynamic_signature(&self) -> Signature<'_>;
}

Types with dynamic signatures.

Prefer implementing Type if possible, but if the actual signature of your type cannot be determined until runtime, you can implement this type to support serialization. You should also implement DynamicDeserialize for deserialization.

Required Methods

fn dynamic_signature(&self) -> Signature<'_>

Get the signature for the implementing type.

See Type::signature for details.