zvariant

Module utils

Constants

ARRAY_ALIGNMENT_DBUS 🔒

ARRAY_SIGNATURE_CHAR

The prefix of ARRAY type signature, as a character. Provided for manual signature creation.

ARRAY_SIGNATURE_STR

The prefix of ARRAY type signature, as a string. Provided for manual signature creation.

DICT_ENTRY_ALIGNMENT_DBUS 🔒

DICT_ENTRY_SIG_END_CHAR

The closing character of DICT_ENTRY type signature. Provided for manual signature creation.

DICT_ENTRY_SIG_END_STR

The closing character of DICT_ENTRY type signature, as a string. Provided for manual signature creation.

DICT_ENTRY_SIG_START_CHAR

The opening character of DICT_ENTRY type signature. Provided for manual signature creation.

DICT_ENTRY_SIG_START_STR

The opening character of DICT_ENTRY type signature, as a string. Provided for manual signature creation.

STRUCT_ALIGNMENT_DBUS 🔒

STRUCT_SIG_END_CHAR

The closing character of STRUCT type signature. Provided for manual signature creation.

STRUCT_SIG_END_STR

The closing character of STRUCT type signature, as a string. Provided for manual signature creation.

STRUCT_SIG_START_CHAR

The opening character of STRUCT type signature. Provided for manual signature creation.

STRUCT_SIG_START_STR

The opening character of STRUCT type signature, as a string. Provided for manual signature creation.

VARIANT_SIGNATURE_CHAR

The VARIANT type signature. Provided for manual signature creation.

VARIANT_SIGNATURE_STR

The VARIANT type signature, as a string. Provided for manual signature creation.

Functions

subslice 🔒

Slice the given slice of bytes safely and return an error if the slice is too small.

usize_to_u8 🔒

usize_to_u32 🔒