Function zvariant::from_slice_fds_for_signature

pub fn from_slice_fds_for_signature<'d, 'r: 'd, B, T: ?Sized>(
    bytes: &'r [u8],
    fds: Option<&[RawFd]>,
    ctxt: EncodingContext<B>,
    signature: &Signature<'_>
) -> Result<T>where
    B: ByteOrder,
    T: Deserialize<'d>,

Deserialize T from a given slice of bytes containing file descriptor indices, with the given signature.

Please note that actual file descriptors are not part of the encoding and need to be transferred via an out-of-band platform specific mechanism. The encoding only contain the indices of the file descriptors and hence the reason, caller must pass a slice of file descriptors.

This function is not available on Windows.