zvariant

Struct Structure

pub struct Structure<'a> { /* private fields */ }

A helper type to wrap structs in Value.

API is provided to convert from, and to tuples.

Implementations

impl<'a> Structure<'a>

pub fn fields(&self) -> &[Value<'a>]

Get a reference to all the fields of self.

pub fn into_fields(self) -> Vec<Value<'a>>

Converts self to a Vec containing all its fields.

pub fn signature(&self) -> Signature<'static>

Get the signature of this Structure.

NB: This method potentially allocates and copies. Use full_signature if you’d like to avoid that.

pub fn full_signature(&self) -> &Signature<'_>

Get the signature of this Structure.

pub fn try_clone(&self) -> Result<Self, Error>

Attempt to clone self.